+++ src/scripts/train/OLMo3/OLMo3-32B-long-context.py
@@ -14,7 +14,6 @@ from olmo_core.internal.experiment import (
     main,
 )
 from olmo_core.launch.beaker import OLMoCoreBeakerImage
-from olmo_core.nn.attention import AttentionBackendName, SlidingWindowAttentionConfig
 from olmo_core.nn.rope import YaRNRoPEScalingConfig
 from olmo_core.nn.transformer import (
     TransformerActivationCheckpointingMode,
@@ -47,18 +46,11 @@ LR = 0.0002071235285  # same as midtraining
 
 
 def build_model_config(common: CommonComponents) -> TransformerConfig:
-    config = TransformerConfig.olmo2_32B(vocab_size=common.tokenizer.padded_vocab_size())
-    config.block.attention.sliding_window = SlidingWindowAttentionConfig(
-        force_full_attention_on_first_layer=False,
-        force_full_attention_on_last_layer=True,
-        pattern=[4096, 4096, 4096, -1],
-    )
-    config.block.attention.backend = AttentionBackendName.flash_2
-
-    config = config.with_rope_scaling(
+    return TransformerConfig.olmo3_32B(
+        vocab_size=common.tokenizer.padded_vocab_size(),
+    ).with_rope_scaling(
         YaRNRoPEScalingConfig(factor=8, beta_fast=32, beta_slow=1, old_context_len=8192)
     )
-    return config
 
 
 def build_train_module_config(common: CommonComponents) -> TransformerTrainModuleConfig: